Modernise to async/await & update deps
This is effectively a rewrite to use async/await, real promises (to
remove the dependency on `q`), `got` instead of `request`, `jest`
instead of mocha with expect.js.

This removes a bunch of security vulnerabilities from outdated
dependencies, but also makes the code a bit simpler.

This does mean node prior to v7.6 isn't supported, but that doesn't seem
like much of an issue.